// src/generated/types.ts
export type CharacterOrderField = "CREATED_AT" | "FAVORITE_CHARACTERS_COUNT";
// ...
export type EpisodeOrderField = "CREATED_AT" | "SORT_NUMBER";
// ...
export type OrderDirection = "ASC" | "DESC";
// ...
export type OrganizationOrderField = "CREATED_AT" | "FAVORITE_ORGANIZATIONS_COUNT";
// ...
export type PersonOrderField = "CREATED_AT" | "FAVORITE_PEOPLE_COUNT";
// ...
export type RatingState = "AVERAGE" | "BAD" | "GOOD" | "GREAT";
// ...
export type StatusState =
  | "NO_STATE"
  | "ON_HOLD"
  | "STOP_WATCHING"
  | "WANNA_WATCH"
  | "WATCHED"
  | "WATCHING";
// ...
export type WorkOrderField = "CREATED_AT" | "SEASON" | "WATCHERS_COUNT";